started separating event algos into folders
